requirements.txt
FastAPI
uvicorn[standard]
sqlalchemy
argon2-cffi
